moved defines
